+++ projects/angular-rave/src/public-api.ts
@@ -5,5 +5,5 @@
 export * from './lib/angular-rave.service';
 export * from './lib/angular-rave.component';
 export * from './lib/angular-rave.directive';
-export { RaveOptions } from './lib/rave-options';
+export { RaveOptions, PaymentInstance } from './lib/rave-options';
 export * from './lib/angular-rave.module';