+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { RaveOptions } from 'angular-rave';
+import { RaveOptions, PaymentInstance } from 'angular-rave';
 
 @Component({
   selector: 'app-root',
@@ -8,11 +8,12 @@ import { RaveOptions } from 'angular-rave';
 })
 export class AppComponent {
   title = 'app';
-  paymentInstance: any;
+  paymentInstance: PaymentInstance;
   raveOptions: RaveOptions = {
     customer_email: 'user@ravemail.com',
     customer_phone: '090848484843',
     amount: 3000,
+    redirect_url: 'http://localhost:4200',
     currency: 'NGN',
     custom_title: 'Angular rave test payment',
     txref: `${Math.random() * 1000000}`,
@@ -20,6 +21,7 @@ export class AppComponent {
 
   paymentFailure() {
     console.log('Payment Failed');
+    this.paymentInstance.close();
   }
 
   paymentSuccess(res) {
@@ -29,6 +31,7 @@ export class AppComponent {
 
   paymentInit(paymentInstance) {
     this.paymentFailure = paymentInstance;
+    console.log(paymentInstance);
     console.log('Payment about to begin', paymentInstance);
   }
 }